Tweaking bootloader compilation.
It now reuses the same vusb sources as the main firmware.
It now compiles for ATmega8.

There is a warning that I haven't managed to fix:
../usbdrv/usbdrv.h:223:24: warning: ‘usbFunctionDescriptor’ used but never defined
I've also tried compiling my other AVR project (from 2011), and it now gives the
same warning. I believe something in the toolchain must have changed.

I've tried using the bootloader from this commit. It does not work.
It gets recognized by the kernel, but avrdude can't find it.